bugprone-implicit-widening-of-multiplication-result fixes for bookKeeping.C and exchangeAMRDonors.C

+++ src/bookKeeping.C
@@ -42,7 +42,7 @@ void MeshBlock::getDonorPacket(PACKET* sndPack, int nsend) const
     //
     for (i = 0; i < nsearch; i++) {
         if (donorId[i] > -1) {
-            k = isearch[2 * i];
+            k = isearch[static_cast<int>(2 * i)];
             sndPack[k].nints += 3;
             sndPack[k].nreals++;
         }
@@ -59,7 +59,7 @@ void MeshBlock::getDonorPacket(PACKET* sndPack, int nsend) const
     //
     for (i = 0; i < nsearch; i++) {
         if (donorId[i] > -1) {
-            k = isearch[2 * i];
+            k = isearch[static_cast<int>(2 * i)];
             sndPack[k].intData[icount[k]++] = meshtag;            // mesh tag
             sndPack[k].intData[icount[k]++] = isearch[2 * i + 1]; // point id
             sndPack[k].intData[icount[k]++] = i; // point id on the donor side
@@ -76,7 +76,7 @@ void MeshBlock::getMBDonorPktSizes(
 {
     for (int i = 0; i < nsearch; i++) {
         if (donorId[i] > -1) {
-            int const ii = isearch[3 * i];
+            int const ii = isearch[static_cast<int>(3 * i)];
             nints[ii] += 4;
             nreals[ii] += 2;
         }
@@ -93,7 +93,7 @@ void MeshBlock::getMBDonorPackets(
             continue;
         }
 
-        int const k = isearch[3 * i];
+        int const k = isearch[static_cast<int>(3 * i)];
         int& ix = ixOffset[k];
         int& rx = rxOffset[k];
 
@@ -180,7 +180,7 @@ void MeshBlock::processDonors(
         if (verbose != 0) TRACEI(iblank[i]);
         if (verbose != 0) TRACED(nodeRes[i]);
         if (verbose != 0) {
-            printf("%f %f %f\n", x[3 * i], x[3 * i + 1], x[3 * i + 2]);
+            printf("%f %f %f\n", x[static_cast<int>(3 * i)], x[3 * i + 1], x[3 * i + 2]);
         }
         if (donorList[i] == nullptr) {
             if (verbose != 0) {
@@ -189,7 +189,7 @@ void MeshBlock::processDonors(
             for (j = 0; j < nmesh; j++) {
                 if (j != (meshtag - BASE) && (holemap[j].existWall != 0)) {
                     if (checkHoleMap(
-                            &x[3 * i], holemap[j].nx, holemap[j].sam,
+                            &x[static_cast<int>(3 * i)], holemap[j].nx, holemap[j].sam,
                             holemap[j].extents) != 0) {
                         iblank[i] = 0;
                         break;
@@ -218,7 +218,7 @@ void MeshBlock::processDonors(
                 if (j != (meshtag - BASE) && (holemap[j].existWall != 0)) {
                     if (iflag[j] == 0) {
                         if (checkHoleMap(
-                                &x[3 * i], holemap[j].nx, holemap[j].sam,
+                                &x[static_cast<int>(3 * i)], holemap[j].nx, holemap[j].sam,
                                 holemap[j].extents) != 0) {
                             iblank[i] = 0;
                             break;
@@ -241,7 +241,7 @@ void MeshBlock::processDonors(
                     myid, wbcnode[i] - BASE, donorList[wbcnode[i] - BASE]);
                 ii = wbcnode[i] - BASE;
                 printf(
-                    "xloc=%e %e %e\n", x[3 * ii], x[3 * ii + 1], x[3 * ii + 2]);
+                    "xloc=%e %e %e\n", x[static_cast<int>(3 * ii)], x[3 * ii + 1], x[3 * ii + 2]);
                 printf(
                     "Computations will continue, but may suffer from accuracy "
                     "problems\n");
@@ -406,7 +406,7 @@ void MeshBlock::processDonors(
             for (j = 0; j < nmesh; j++) {
                 if (j != (meshtag - BASE) && (holemap[j].existWall != 0U)) {
                     int const SB_val =
-                        checkAdaptiveHoleMap(&x[3 * i], &holemap[j]);
+                        checkAdaptiveHoleMap(&x[static_cast<int>(3 * i)], &holemap[j]);
                     if (SB_val != OUTSIDE_SB) {
                         iblank[i] = 0;
                         break;
@@ -444,7 +444,7 @@ void MeshBlock::processDonors(
                         // body{j} does NOT have candidate so check if point is
                         // INSIDE SB
                         int const SB_val =
-                            checkAdaptiveHoleMap(&x[3 * i], &holemap[j]);
+                            checkAdaptiveHoleMap(&x[static_cast<int>(3 * i)], &holemap[j]);
                         if (SB_val != OUTSIDE_SB) {
                             iblank[i] = 0;
                             break;
@@ -469,7 +469,7 @@ void MeshBlock::processDonors(
                     myid, wbcnode[i] - BASE, donorList[wbcnode[i] - BASE]);
                 ii = wbcnode[i] - BASE;
                 fprintf(
-                    stderr, "xloc=%e %e %e\n", x[3 * ii], x[3 * ii + 1],
+                    stderr, "xloc=%e %e %e\n", x[static_cast<int>(3 * ii)], x[3 * ii + 1],
                     x[3 * ii + 2]);
                 fprintf(
                     stderr,
@@ -650,7 +650,7 @@ void MeshBlock::findInterpData(int* recid, int irecord, double receptorRes2)
     // if (myid==63 && irecord==3224) verbose=1;
     // if (myid==1 && irecord==158192 && meshtag==1) verbose=1;
     receptorRes = fabs(receptorRes2);
-    procid = isearch[3 * irecord];
+    procid = isearch[static_cast<int>(3 * irecord)];
     pointid = isearch[3 * irecord + 1];
     blockid = isearch[3 * irecord + 2];
     meshtagrecv = tagsearch[irecord];
